Update assignmentII.py
Write a Python program to calculate the roots of a quadratic equation using the quadratic formula. The program prompts the user to enter the coefficients A, B and C and calculates the two roots. If the roots are equal only one root is displayed. Hint:
Root1 = (-B + B  4  A  C
2
) / (2.0 A)
Root2 = (-B - B  4  A  C
2
) / (2.0 A)